+++ src/main/java/com/l299l/newbedwars/arena/team/Team.java
@@ -56,6 +56,8 @@ public class Team {
             player.setInvulnerable(false);
             player.setGameMode(GameMode.SURVIVAL);
             player.setBedSpawnLocation(teamSpawn, true);
+            player.getEnderChest().clear();
+            //TODO: add restoring normal world ender chest
         }
         teamBed.respawn();
         teamShop.start();